Moved UserInfoView to separate file
